% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain_plot.R
\name{mountain_plot}
\alias{mountain_plot}
\title{Create a Mountain Plot}
\usage{
mountain_plot(x, y, plot.type)
}
\arguments{
\item{x}{First input numeric vector}

\item{y}{Second input numeric vector}

\item{plot.type}{The plot type to use. Defaults to none. Options include 'plotly'.}
}
\value{
A plot object
}
\description{
This function creates a mountain plot. It takes two numeric vectors x and y as inputs.
It also takes an optional input to specify if the plotly package should
be used or not for plotting.
}
